% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survbayes-package.R
\name{concatenate_formula}
\alias{concatenate_formula}
\title{Concatenate a list of formula}
\usage{
concatenate_formula(formula, forms, data)
}
\arguments{
\item{formula}{Main formula}

\item{forms}{List of rhs-only formula, from `anc` argument.}

\item{data}{Data.frame}
}
\value{
Formula
}
\description{
Concatenate a list of formula
}
