% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survbayes-package.R
\name{contr.full}
\alias{contr.full}
\title{Contrasts where each factor-level is explicitly represented}
\usage{
contr.full(n, contrasts = TRUE, sparse = FALSE)
}
\arguments{
\item{n}{a vector of levels for a factor, or the number of levels.}

\item{contrasts}{Just set this to TRUE}

\item{sparse}{logical indicating if the result should be sparse}
}
\value{
Contrast-matrix
}
\description{
This contrast-type should only be used in \code{survreg_map}, where (a) the function knows to
center the contrasts so that each coefficient estimate can be interpreted as "deviation from
average," and (b) the fact that this contrast-type is linearly dependent isn't a problem, thanks
to shrinkage from priors.
}
