% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survbayes-package.R
\name{fit_survreg_map}
\alias{fit_survreg_map}
\title{Fit a bayesian survival model using MAP}
\usage{
fit_survreg_map(survreg_data, newdata = NULL)
}
\arguments{
\item{survreg_data}{An object from \code{make_survreg_data}}
}
\value{
An object of class \code{survreg_map}, with print/plot/predict/logLik methods.
}
\description{
This function usually isn't called directly. Instead, you'd call \code{survreg_map}, which will
call \code{make_survreg_data}, then call this function.
}
