% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survbayes-package.R
\name{get_dist_info}
\alias{get_dist_info}
\title{Get information about a survival distribution}
\usage{
get_dist_info(dist, k = NULL)
}
\arguments{
\item{dist}{Character naming a distribution}

\item{k}{For spline-based distributions, the number of knots}
}
\value{
A list with information, functions, etc.
}
\description{
Get information about a survival distribution
}
