% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roy-parm-distribution.R
\name{get_roy_parm_function_block_code}
\alias{get_roy_parm_function_block_code}
\title{Get Stan code for helper functions for Royston/Parmar-splines}
\usage{
get_roy_parm_function_block_code()
}
\value{
a character string of model code for the stan language
}
\description{
Get Stan code for helper functions for Royston/Parmar-splines
}
