% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survbayes-package.R
\name{get_terms_mapping}
\alias{get_terms_mapping}
\title{Get mapping from column names to model-terms and vice versa}
\usage{
get_terms_mapping(formula, data, ...)
}
\arguments{
\item{formula}{The formula to be passed to model.frame}

\item{data}{The data.frame}

\item{...}{Arguments to be passed to \code{stats::model.matrix}}
}
\value{
A nested list, where the first level is 'from', the second level is 'to'.
}
\description{
Get mapping from column names to model-terms and vice versa
}
