% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survreg_map.R
\name{logLik.survreg_map}
\alias{logLik.survreg_map}
\title{Method for extracting logLiklihood from \code{survreg_map}.}
\usage{
\method{logLik}{survreg_map}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{Object of class \code{survreg_map}}

\item{newdata}{New data to compute the log-liklihood on.}

\item{...}{Ignored}
}
\value{
Log-liklihood
}
\description{
Method for extracting logLiklihood from \code{survreg_map}.
}
