% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roy-parm-distribution.R
\name{make_roy_parm_spline_stan_model_code}
\alias{make_roy_parm_spline_stan_model_code}
\title{Generate Stan model-code for fitting a model with the Royston/Parmar-spline-survival distribution}
\usage{
make_roy_parm_spline_stan_model_code(model_mats, priors)
}
\arguments{
\item{model_mats}{List of model-matrices}

\item{priors}{A dataframe specifying priors}
}
\value{
a character string of model code for the stan language
}
\description{
Generate Stan model-code for fitting a model with the Royston/Parmar-spline-survival distribution
}
