% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survbayes-package.R
\name{merge_formulae}
\alias{merge_formulae}
\title{Merge a list of formulae}
\usage{
merge_formulae(forms, data)
}
\arguments{
\item{forms}{List of formulae/formulas}

\item{data}{Data.frame}
}
\value{
Single formula, with environment(forms[[1]])
}
\description{
Merge a list of formulae
}
