% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survbayes-package.R
\name{prep_model_components}
\alias{prep_model_components}
\title{Prepare components of a model (model.frame, response, etc) for bayesian survival-regression}
\usage{
prep_model_components(formula_concat, forms, data, na.action, xlev,
  standardize_y)
}
\arguments{
\item{formula_concat}{A formula containing all the terms for all the parameters}

\item{forms}{A list of formula, one for each parameter}

\item{data}{The data.frame}

\item{na.action}{Function for NAs}

\item{xlev}{Passed to model.frame}

\item{standardize_y}{If TRUE, then response variable is standardized. This means taking the log
of the 'end' times, standardizing (centering/scaling) them, then taking the exponent of that.
If this is a list, it's assumed to have components 'center' and 'scale', for performing this
standardization.}
}
\value{
A list with needed components, such as model.frame/mats, response object, and info for
  subsequent calls (params for standardizing).
}
\description{
Prepare components of a model (model.frame, response, etc) for bayesian survival-regression
}
