% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survbayes-package.R
\name{prep_survreg_data}
\alias{prep_survreg_data}
\title{Prepare data for a bayesian survival-regression model}
\usage{
prep_survreg_data(formula, anc = NULL, data,
  distribution = "roy_parm_splines", dist_config = list(knots = NULL),
  standardize_y = TRUE, na.action = na.exclude)
}
\arguments{
\item{formula}{Formula with both rhs and lhs. Rhs will be applied to 'location' parameter of the distribution, or for spline-based models, the first parameter (gamma0).}

\item{anc}{A list of lhs-only formulae. Must be named for the parameters of the distribution (e.g., \code{list(gamma1 = ~ pred1 + pred2)}).}

\item{data}{A data.frame}

\item{distribution}{Character naming the distribution}

\item{dist_config}{A list customizing the distribution. For example, if \code{distribution = 'roy_parm_splines'}, then you can pass \code{list(knots=)} to specify the location of the knots.}

\item{standardize_y}{If TRUE (the default), then response variable is standardized. This means taking the log
of the 'end' times, standardizing (centering/scaling) them, then taking the exponent of that.
If this is a list, it's assumed to have components 'center' and 'scale', for performing this
standardization. Standardization is helpful because it makes the default priors more meaningful
across datasets.}

\item{na.action}{Function for NAs}
}
\value{
A list with components that can be passed to \code{rstan::stan}, or the list can be passed to \code{survreg_map}.
}
\description{
Prepare data for a bayesian survival-regression model
}
