% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survreg_stan.R
\name{stan_to_r_pdf}
\alias{stan_to_r_pdf}
\title{Map the name of a PDF in Stan to its name in R}
\usage{
stan_to_r_pdf(stan_pdf)
}
\arguments{
\item{stan_pdf}{Chraracter-string with stan name}
}
\value{
Character-string with R function name
}
\description{
Map the name of a PDF in Stan to its name in R
}
