% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survbayes-package.R
\name{standardize_formulae}
\alias{standardize_formulae}
\title{Standardize formula arguments}
\usage{
standardize_formulae(formula, anc, dist_info)
}
\arguments{
\item{formula}{The main formula}

\item{anc}{List of others}

\item{dist_info}{Information about the distribution}
}
\value{
A list of formulas, now matched to the distribution
}
\description{
Standardize formula arguments
}
