% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survbayes-package.R
\name{Survint}
\alias{Survint}
\title{Survival object allowing for data that is both interval-censored *and* truncated}
\usage{
Survint(end, event, start = NULL, end_lb = NULL)
}
\arguments{
\item{end}{Last observed time.}

\item{event}{Event indicator. When `event==0`, `end_lb` is ignored, and `end` is considered
right-censored. When `event==1` and `end > end_lb`, the event is considered interval-censored
(occurring sometime between these two times); when `event==1` and `end == end_lb`, the `end` is
considered an exact time of event.}

\item{start}{Truncation time. If unspecified}

\item{end_lb}{Optional. Lower bound of last-observed time. If included, then an `event`
indicator signals inteval-censoring (i.e., an event some time between `end_lb` and `end`). If
some observations are interval-censored, while for some the exact time is known, the latter can
be indicated by rows where `end == end_lb`. If the `end_lb` argument is omitted entirely, then
`end == end_lb` is assumed for all data.}
}
\value{
An object of class 'Survint', which can be used as the response-object in
  \code{survreg_stan} or \code{survreg_map}.
}
\description{
Survival object allowing for data that is both interval-censored *and* truncated
}
