% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survbayes-package.R
\name{fit_survreg_map}
\alias{fit_survreg_map}
\title{Fit a bayesian survival model using MAP}
\usage{
fit_survreg_map(survreg_data, newdata = NULL, optim_args = list(method =
  "BFGS", control = list(trace = as.integer(interactive()), maxit = 250)))
}
\arguments{
\item{survreg_data}{An object from \code{make_survreg_data}}

\item{newdata}{New data to use in fitting}

\item{optim_args}{A list of arguments to be passed to \code{optim}.}
}
\value{
An object of class \code{survreg_map}, with print/plot/predict/logLik methods.
}
\description{
This function usually isn't called directly. Instead, you'd call \code{survreg_map}, which will
call \code{make_survreg_data}, then call this function.
}
