% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survbayes-package.R
\name{logLik.survreg_map}
\alias{logLik.survreg_map}
\title{Get logliklihood from an object of class \code{survreg_map}.}
\usage{
\method{logLik}{survreg_map}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{An object of class \code{survreg_map}.}

\item{newdata}{Optional, newdata on which to calculate the likelihood.}

\item{...}{Ignored}
}
\value{
The log-likelihood.
}
\description{
Get logliklihood from an object of class \code{survreg_map}.
}
