% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survbayes-package.R
\name{plot_coefs}
\alias{plot_coefs}
\title{Plot model coefficients}
\usage{
plot_coefs(object, ...)
}
\arguments{
\item{object}{A model object}

\item{...}{Other arguments to be passed to methods}
}
\value{
A ggplot object
}
\description{
Plot model coefficients
}
