% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survbayes-package.R
\name{set_prior}
\alias{set_prior}
\title{Set prior}
\usage{
set_prior(parameter = NULL, terms, family = NULL, location = NULL,
  spread = NULL, fixed = NULL)
}
\arguments{
\item{parameter}{The parameter}

\item{terms}{The terms. Can be a character vector, or the result of \code{dplyr::vars}, which can
be convenient for setting the prior on all but some terms.}

\item{family}{The family for the prior distribution}

\item{location}{A value for location on the distribution.}

\item{spread}{A value for the spread (scale) on the distribution}

\item{fixed}{In the fitting process, should this parameter be fixed at its prior value, or updated?}
}
\value{
An object of class 'prior', to be (optionally inserted into a list with other priors and)
  passed to `priors` arg in `survreg_map`
}
\description{
Set prior
}
