% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survbayes-package.R
\name{update.survreg_map}
\alias{update.survreg_map}
\title{Update method for \code{survreg_map}}
\usage{
\method{update}{survreg_map}(object, ...)
}
\arguments{
\item{object}{An object of class \code{survreg_map}}

\item{...}{Args to be updated. Currently just `data` and `priors`. Even if `data` is updated, the
centering/scaling of predictors will *not* be updated, so that the updated model can be
compared to the original apples-to-apples.}
}
\value{
An updated model.
}
\description{
This is a rudimentary update method for \code{survreg_map}. Unlike most update methods in R, this
doesn't work by modifying and replaying the call, but instead uses the slots in the object. It's
limited to updating the formula(s), the data, and/or the prior(s).
}
